requirements.txt
# Core AI and LangChain
langchain>=0.3.0
langchain-openai>=0.2.0

# OpenAI API client
openai>=1.50.0

# Web service (REST API)
fastapi>=0.111.0
uvicorn>=0.30.0

# Environment variables
python-dotenv>=1.0.1

# Data handling (nếu bạn muốn load schema, JSON, v.v.)
pydantic>=2.8.0

# Optional: logging, debugging
rich>=13.7.0

# Optional: HTTP client (gọi API nội bộ)
httpx>=0.27.0